requirements-doc.txt
doctr
doctr-versions-menu
graphviz
ipython<7
matplotlib
numpydoc
sphinx>=3.2.1
sphinx_rtd_theme>=0.5.0
